feat: Add product management components including modals and tabs for managing product relationships, details, and settings
- Implemented ProductRelationshipsModal for managing substitutes and similar products.
- Created ProductDetailModal to display detailed information about a selected product.
- Developed ProductDetailsTab for editing product information including description, safety advice, and dosage form.
- Added SettingsTab for configuring product settings such as schedule, tax rate, and storage conditions.
- Introduced StatsCard component for displaying product statistics.
- Populated product catalog data with sample products and categories.
- Defined product and product form types for better type safety and structure.
