+++ src/app/admin/dashboard/pharmacy/product/components/ProductDetailModal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Product } from "@/app/types/product";
+import { Product } from "../types/product";
 
 interface ProductDetailModalProps {
   isOpen: boolean;
