+++ src/app/admin/dashboard/pharmacy/product/components/StatsCard.tsx
