+++ src/app/admin/dashboard/pharmacy/product/data/productCatalogData.ts
@@ -1,0 +1,229 @@
+import { Product } from "../types/product";
+
+export const products: Product[] = [
+  {
+    id: "1",
+    name: "Paracetamol 500mg",
+    code: "MED-001",
+    category: "Medicines",
+    subcategory: "Healthcare Pharma",
+    mrp: 120.0,
+    sellingPrice: 99.0,
+    discount: 17.5,
+    stock: 1200,
+    status: "Active",
+    featured: true,
+    substitutes: 2,
+    similar: 1,
+    brand: "",
+    composition: "",
+    dosageForm: "",
+    strength: "",
+    packSize: "",
+    schedule: "",
+    taxRate: 0,
+    hsnCode: "",
+    storageConditions: "",
+    shelfLife: 0,
+    prescriptionRequired: false,
+    saftyDescription: "",
+    storageDescription: "",
+    createdAt: "",
+    similarCount: 0,
+    description: "",
+    substitutesCount: 0,
+    manufacturer: "",
+  },
+  {
+    id: "2",
+    name: "Ibuprofen 400mg",
+    code: "MED-002",
+    category: "Medicines",
+    subcategory: "Healthcare Pharma",
+    mrp: 150.0,
+    sellingPrice: 120.0,
+    discount: 20,
+    stock: 1500,
+    status: "Active",
+    featured: true,
+    substitutes: 3,
+    similar: 2,
+    brand: "",
+    composition: "",
+    dosageForm: "",
+    strength: "",
+    packSize: "",
+    schedule: "",
+    taxRate: 0,
+    hsnCode: "",
+    storageConditions: "",
+    shelfLife: 0,
+    prescriptionRequired: false,
+    saftyDescription: "",
+    storageDescription: "",
+    createdAt: "",
+    similarCount: 0,
+    description: "",
+    substitutesCount: 0,
+    manufacturer: "",
+  },
+  {
+    id: "3",
+    name: "Aspirin 325mg",
+    code: "MED-003",
+    category: "Medicines",
+    subcategory: "Healthcare Pharma",
+    mrp: 80.0,
+    sellingPrice: 60.0,
+    discount: 25,
+    stock: 45,
+    status: "Active",
+    featured: false,
+    substitutes: 1,
+    similar: 1,
+    brand: "",
+    composition: "",
+    dosageForm: "",
+    strength: "",
+    packSize: "",
+    schedule: "",
+    taxRate: 0,
+    hsnCode: "",
+    storageConditions: "",
+    shelfLife: 0,
+    prescriptionRequired: false,
+    saftyDescription: "",
+    storageDescription: "",
+    createdAt: "",
+    similarCount: 0,
+    description: "",
+    substitutesCount: 0,
+    manufacturer: "",
+  },
+  {
+    id: "4",
+    name: "Naproxen 250mg",
+    code: "MED-004",
+    category: "Medicines",
+    subcategory: "Healthcare Pharma",
+    mrp: 200.0,
+    sellingPrice: 160.0,
+    discount: 20,
+    stock: 0,
+    status: "Inactive",
+    featured: false,
+    substitutes: 3,
+    similar: 2,
+    brand: "",
+    composition: "",
+    dosageForm: "",
+    strength: "",
+    packSize: "",
+    schedule: "",
+    taxRate: 0,
+    hsnCode: "",
+    storageConditions: "",
+    shelfLife: 0,
+    prescriptionRequired: false,
+    saftyDescription: "",
+    storageDescription: "",
+    createdAt: "",
+    similarCount: 0,
+    description: "",
+    substitutesCount: 0,
+    manufacturer: "",
+  },
+  {
+    id: "5",
+    name: "Digital Thermometer",
+    code: "DEV-001",
+    category: "Medical Devices",
+    subcategory: "Diagnostic Equipment",
+    mrp: 200.0,
+    sellingPrice: 160.0,
+    discount: 20,
+    stock: 890,
+    status: "Active",
+    featured: false,
+    substitutes: 3,
+    similar: 2,
+    brand: "",
+    composition: "",
+    dosageForm: "",
+    strength: "",
+    packSize: "",
+    schedule: "",
+    taxRate: 0,
+    hsnCode: "",
+    storageConditions: "",
+    shelfLife: 0,
+    prescriptionRequired: false,
+    saftyDescription: "",
+    storageDescription: "",
+    createdAt: "",
+    similarCount: 0,
+    description: "",
+    substitutesCount: 0,
+    manufacturer: "",
+  },
+  {
+    id: "6",
+    name: "Vitamin D3 Tablets",
+    code: "SUP-001",
+    category: "Supplements",
+    subcategory: "Vitamins",
+    mrp: 200.0,
+    sellingPrice: 160.0,
+    discount: 20,
+    stock: 100,
+    status: "Active",
+    featured: true,
+    substitutes: 3,
+    similar: 2,
+    brand: "",
+    composition: "",
+    dosageForm: "",
+    strength: "",
+    packSize: "",
+    schedule: "",
+    taxRate: 0,
+    hsnCode: "",
+    storageConditions: "",
+    shelfLife: 0,
+    prescriptionRequired: false,
+    saftyDescription: "",
+    storageDescription: "",
+    createdAt: "",
+    similarCount: 0,
+    description: "",
+    substitutesCount: 0,
+    manufacturer: "",
+  },
+];
+
+export const categories = [
+  "All Categories",
+  "Medicines",
+  "Medical Devices",
+  "Supplements",
+  "Personal Care",
+];
+
+export const sortOptions = [
+  "Sort",
+  "Relevance (default)",
+  "Selling Price - Low to High",
+  "Selling Price - High to Low",
+  "Product Status",
+  "By Name",
+  "By Stock",
+  "Discount",
+];
+
+export const tabs = [
+  "All Products",
+  "Active",
+  "Featured",
+  "Out of Stock",
+  "Inactive",
+];
