+++ src/app/admin/dashboard/pharmacy/product/types/product.ts
