+++ src/app/admin/dashboard/pharmacy/product/types/productForm.type.ts
